requirements.txt
pandas==1.4.2
pycryptodomex==3.15.0
pyperclip==1.8.2
python-dotenv==0.21.0
SQLAlchemy==1.4.41
